Moved StateTag interface to its rightful place in package automata.
