+++ src/io/github/sboyanovich/scannergenerator/scanner/DomainTag.java
@@ -1,5 +1,6 @@
 package io.github.sboyanovich.scannergenerator.scanner;
 
+import io.github.sboyanovich.scannergenerator.automata.StateTag;
 import io.github.sboyanovich.scannergenerator.scanner.token.Domain;
 
 /**
