+++ src/io/github/sboyanovich/scannergenerator/scanner/Scanner.java
@@ -1,5 +1,6 @@
 package io.github.sboyanovich.scannergenerator.scanner;
 
+import io.github.sboyanovich.scannergenerator.automata.StateTag;
 import io.github.sboyanovich.scannergenerator.scanner.token.Domain;
 import io.github.sboyanovich.scannergenerator.scanner.token.Token;
 import io.github.sboyanovich.scannergenerator.utility.Utility;
