+++ src/io/github/sboyanovich/scannergenerator/tests/biglexgen/LexGenScanner.java
@@ -2,6 +2,7 @@ package io.github.sboyanovich.scannergenerator.tests.biglexgen;
 
 import io.github.sboyanovich.scannergenerator.automata.DFA;
 import io.github.sboyanovich.scannergenerator.automata.NFA;
+import io.github.sboyanovich.scannergenerator.automata.StateTag;
 import io.github.sboyanovich.scannergenerator.scanner.*;
 import io.github.sboyanovich.scannergenerator.scanner.token.Domain;
 import io.github.sboyanovich.scannergenerator.scanner.token.Token;
