+++ src/io/github/sboyanovich/scannergenerator/tests/biglexgen/RecognizerGenTest.java
@@ -5,7 +5,7 @@ import io.github.sboyanovich.scannergenerator.automata.NFA;
 import io.github.sboyanovich.scannergenerator.scanner.LexicalRecognizer;
 import io.github.sboyanovich.scannergenerator.scanner.Message;
 import io.github.sboyanovich.scannergenerator.scanner.Position;
-import io.github.sboyanovich.scannergenerator.scanner.StateTag;
+import io.github.sboyanovich.scannergenerator.automata.StateTag;
 import io.github.sboyanovich.scannergenerator.scanner.token.Domain;
 import io.github.sboyanovich.scannergenerator.scanner.token.Token;
 import io.github.sboyanovich.scannergenerator.utility.Utility;
