+++ src/io/github/sboyanovich/scannergenerator/tests/mockjava/MockJavaTest.java
@@ -3,6 +3,7 @@ package io.github.sboyanovich.scannergenerator.tests.mockjava;
 import io.github.sboyanovich.scannergenerator.automata.DFA;
 import io.github.sboyanovich.scannergenerator.automata.NFA;
 import io.github.sboyanovich.scannergenerator.automata.NFAStateGraphBuilder;
+import io.github.sboyanovich.scannergenerator.automata.StateTag;
 import io.github.sboyanovich.scannergenerator.scanner.Compiler;
 import io.github.sboyanovich.scannergenerator.scanner.*;
 import io.github.sboyanovich.scannergenerator.scanner.Scanner;
